requirements.txt
mongoengine>=0.8.7
Django>=1.6.2
Sphinx>=1.2.2
Twisted>=13.2.0
autobahn>=0.8.7
pymongo>=2.7
requests>=2.2.1
django-tastypie>=0.11.0
django-tastypie-mongoengine>=0.4.5
django-tastypie-swagger>=0.1.2
pytz>=2014.2
pyOpenSSL==0.13
lisa-plugin-ChatterBot
PyWit>=0.2.0
